% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{indexfunds}
\alias{indexfunds}
\title{IndexFunds}
\usage{
indexfunds()
}
\value{
shiny app
}
\description{
It will provide the price of the selected ETF. Already pre-select four tickers of the most promising ETF's to invest.
}
